requirements.txt
pandas>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
tabulate>=0.9.0
colorama>=0.4.6
matplotlib>=3.5.0
seaborn>=0.12.0 